Fixed some tests and export blocks

+++ admin/cloud/layout-bridge.php
@@ -38,9 +38,12 @@ class Brizy_Admin_Cloud_LayoutBridge extends Brizy_Admin_Cloud_AbstractBridge {
 			$bridge->export( $fontUid );
 		}
 
+		$bridge = new Brizy_Admin_Cloud_ScreenshotBridge( $this->client );
+		$bridge->export( $layout );
+
 		$layoutObject = $this->client->createOrUpdateLayout( $layout );
 
-		$layout->setSynchronized( Brizy_Editor_Project::get()->getCloudAccountId(), $layoutObject->uid );
+		$layout->setSynchronized( $this->client->getBrizyProject()->getCloudAccountId(), $layoutObject->uid );
 
 		$layout->saveStorage();
 	}
@@ -52,30 +55,70 @@ class Brizy_Admin_Cloud_LayoutBridge extends Brizy_Admin_Cloud_AbstractBridge {
 	 * @throws Exception
 	 */
 	public function import( $layoutId ) {
-		$layouts = $this->client->getLayouts( [ 'filter' => [ 'uid' => $layoutId ] ] );
+		global $wpdb;
+
+		$layouts = $this->client->getLayouts( [ 'uid' => $layoutId ] );
 
 		if ( ! isset( $layouts[0] ) ) {
 			return;
 		}
 
-		$layout = $layouts[0];
+		try {
+			$wpdb->query( 'START TRANSACTION ' );
 
-		$name = md5( time() );
-		$post = wp_insert_post( array(
-			'post_title'  => $name,
-			'post_name'   => $name,
-			'post_status' => 'publish',
-			'post_type'   => Brizy_Admin_Layouts_Main::CP_LAYOUT
-		) );
+			$layout = (array) $layouts[0];
 
-		if ( $post ) {
-			$brizyPost = Brizy_Editor_Layout::get( $post, $layout['uid'] );
-			$brizyPost->setMeta( $layout['meta'] );
-			$brizyPost->set_editor_data( $layout['data'] );
-			$brizyPost->set_uses_editor( true );
-			$brizyPost->set_needs_compile( true );
-			$brizyPost->saveStorage();
-			$brizyPost->setSynchronized( Brizy_Editor_Project::get()->getCloudAccountId(), $layout['uid'] );
+			$name = md5( time() );
+			$post = wp_insert_post( array(
+				'post_title'  => $name,
+				'post_name'   => $name,
+				'post_status' => 'publish',
+				'post_type'   => Brizy_Admin_Layouts_Main::CP_LAYOUT
+			) );
+
+			if ( $post ) {
+				$brizyPost = Brizy_Editor_Layout::get( $post, $layout['uid'] );
+
+				if ( isset( $layout['media'] ) ) {
+					$brizyPost->setMedia( $layout['media'] );
+				}
+				if ( isset( $layout['meta'] ) ) {
+					$brizyPost->setMeta( $layout['meta'] );
+				}
+				$brizyPost->set_editor_data( $layout['data'] );
+				$brizyPost->set_uses_editor( true );
+				$brizyPost->set_needs_compile( true );
+				$brizyPost->saveStorage();
+				$brizyPost->setDataVersion( 1 );
+				$brizyPost->setSynchronized( $this->client->getBrizyProject()->getCloudAccountId(), $layout['uid'] );
+				$brizyPost->save();
+
+
+				// import fonts
+				if ( isset( $layout['media'] ) ) {
+					$blockMedia = json_decode( $layout['media'] );
+
+					$fontBridge = new Brizy_Admin_Cloud_FontBridge( $this->client );
+					if ( isset( $blockMedia->fonts ) ) {
+						foreach ( $blockMedia->fonts as $cloudFontUid ) {
+							$fontBridge->import( $cloudFontUid );
+						}
+					}
+
+					$mediaBridge = new Brizy_Admin_Cloud_MediaBridge( $this->client );
+					$mediaBridge->setBlockId( $post );
+					if ( isset( $blockMedia->images ) ) {
+						foreach ( $blockMedia->images as $mediaUid ) {
+							$mediaBridge->import( $mediaUid );
+						}
+					}
+				}
+
+			}
+			$wpdb->query( 'COMMIT' );
+		} catch ( Exception $e ) {
+			$wpdb->query( 'ROLLBACK' );
+			Brizy_Logger::instance()->critical( 'Importing layout ' . $layoutId . ' failed', [ $e ] );
 		}
 	}
 
@@ -86,6 +129,9 @@ class Brizy_Admin_Cloud_LayoutBridge extends Brizy_Admin_Cloud_AbstractBridge {
 	 * @throws Exception
 	 */
 	public function delete( $layout ) {
-		$this->client->deleteLayout( $layout->getCloudId() );
+
+		if ( $layout->getCloudId() ) {
+			$this->client->deleteLayout( $layout->getCloudId() );
+		}
 	}
 }
